# isac/control/api/routes_routing.py
"""路由规则与互联 Link 端点 (SPECIFICATION.md 4.4)。

Bearer Token 认证 + 规则持久化 (router/rules.py save_rules) + Link 持久化 (data/links.jsonc) + 审计日志。
"""

from __future__ import annotations

import asyncio
from collections.abc import Callable
from pathlib import Path
from typing import TYPE_CHECKING, Any

from isac.utils.logger import get_logger

if TYPE_CHECKING:
    from isac.control.audit import AuditLog
    from isac.router.router import MessageRouter
    from isac.runtime.bus import InterAgentBus

logger = get_logger(__name__)


def build_router(
    router: MessageRouter,
    bus: InterAgentBus,
    auth_dependency: Any = None,
    audit_log: AuditLog | None = None,
    routing_rules_path: str = "data/routing.jsonc",
    links_path: str = "data/links.jsonc",
    scope_dependency: Any = None,
) -> Any:
    from fastapi import APIRouter, Depends

    from isac.router.rules import save_rules
    from isac.router.types import ChannelBinding, RoutingRules
    from isac.runtime.bus import InterAgentLink

    api = APIRouter(
        tags=["routing"],
        dependencies=[Depends(auth_dependency)] if auth_dependency else [],
    )
    # Fix-12: spec §6.1 只给了 routing:write/link:write, 没有对应的 :read scope;
    # 读操作维持"只要 auth_dependency 认证通过即可", 不额外收窄。
    routing_write_deps = [Depends(scope_dependency("routing:write"))] if scope_dependency else []
    link_write_deps = [Depends(scope_dependency("link:write"))] if scope_dependency else []
    # #29 审计 actor 归因: handler 经此依赖拿真实调用方身份写入审计。
    caller_dep = Depends(auth_dependency) if auth_dependency else Depends(lambda: "anonymous")

    @api.get("/routing/rules")
    async def get_rules() -> dict:
        rules = router.get_rules()
        return {
            "bindings": [vars(b) for b in rules.bindings],
            "default_agents": rules.default_agents,
        }

    @api.put("/routing/rules", dependencies=routing_write_deps)
    async def put_rules(body: dict, caller: str = caller_dep) -> dict:
        rules = RoutingRules(
            bindings=[ChannelBinding(**b) for b in body.get("bindings", [])],
            default_agents=dict(body.get("default_agents", {})),
        )
        router.set_rules(rules)
        # N5 (2026-10-11): 原子写 (tmp+fsync+replace) 迁 to_thread, 不阻塞事件循环。
        await asyncio.to_thread(save_rules, Path(routing_rules_path), rules)
        if audit_log is not None:
            await audit_log.record(
                actor=caller,
                method="PUT",
                path="/api/v1/routing/rules",
                action="update_routing_rules",
                detail=f"{len(rules.bindings)} bindings",
                status_code=200,
            )
        return {"status": "updated"}

    @api.get("/links")
    async def list_links() -> list[dict]:
        return [vars(link) for link in bus.list_links()]

    @api.post("/links", dependencies=link_write_deps)
    async def add_link(body: dict, caller: str = caller_dep) -> dict:
        from fastapi import HTTPException

        try:
            link = InterAgentLink(**body)
        except (ValueError, TypeError) as exc:
            # Fix-16: from_agent/to_agent 格式非法 (含 XSS payload) 在构造期就被
            # InterAgentLink.__post_init__ 拒绝; 转 400 而不是让 500 泄露内部异常。
            # R14: 服务端记录完整异常, 客户端只返回通用错误码 (不泄露内部信息)。
            logger.warning("Link 配置校验失败", error=str(exc), exc_info=True)
            raise HTTPException(
                status_code=400,
                detail={"code": "INVALID_CONFIG", "message": "Link config validation failed"},
            ) from exc
        # add_link 内部已触发 _trigger_persist; 但 routes_routing 持有独立的
        # _persist_links 路径, 用它把磁盘写入错误回传 500 (in-memory 状态已变更,
        # 调用方需要知道不一致) (CODE_REVIEW_REPORT.md #20)。
        bus.add_link(link)
        await _persist_links_or_raise(bus, Path(links_path))
        await _audit_link_change(
            audit_log, method="POST", path="/api/v1/links", action="add_link",
            target=f"{link.from_agent}->{link.to_agent}", actor=caller,
        )
        return {"status": "added"}

    @api.delete("/links", dependencies=link_write_deps)
    async def remove_link(from_agent: str, to_agent: str, caller: str = caller_dep) -> dict:
        bus.remove_link(from_agent, to_agent)
        await _persist_links_or_raise(bus, Path(links_path))
        await _audit_link_change(
            audit_log, method="DELETE", path="/api/v1/links", action="remove_link",
            target=f"{from_agent}->{to_agent}", actor=caller,
        )
        return {"status": "removed"}

    return api


async def _persist_links_or_raise(bus: InterAgentBus, path: Path) -> None:
    """持久化失败抛 HTTPException(500), 让 API 层把磁盘/内存不一致暴露给调用方
    (CODE_REVIEW_REPORT.md #20)。N5 (2026-10-11): 原子写含 fsync, 迁 to_thread。"""
    from fastapi import HTTPException

    def _write() -> None:
        _persist_links(bus, path)

    try:
        await asyncio.to_thread(_write)
    except Exception as exc:
        # R14: 服务端记录完整异常 (含磁盘 IO 错误细节), 客户端只返回通用错误码。
        logger.error("Link 持久化失败", error=str(exc), exc_info=True)
        raise HTTPException(
            status_code=500,
            detail={"code": "LINK_PERSIST_FAILED", "message": "Failed to persist links"},
        ) from exc


async def _audit_link_change(
    audit_log: AuditLog | None,
    *,
    method: str,
    path: str,
    action: str,
    target: str,
    actor: str = "authenticated",
) -> None:
    """Link 变更的统一审计记录 (audit_log 为 None 时跳过)。actor 归因 (#29)。"""
    if audit_log is None:
        return
    await audit_log.record(
        actor=actor,
        method=method,
        path=path,
        action=action,
        target=target,
        status_code=200,
    )


def _persist_links(bus: InterAgentBus, path: Path) -> None:
    """把所有 Link 持久化到 data/links.jsonc (原子替换, K4)。

    失败时抛异常给调用方, 由 API 层返回 500 (CODE_REVIEW_REPORT.md #20)。
    """
    links = [vars(link) for link in bus.list_links()]
    from isac.utils.fs import atomic_write_json

    atomic_write_json(path, {"links": links})


def make_links_persist_callback(bus: InterAgentBus, path: Path) -> Callable[[], None]:
    """构造 bus.set_persist 用的持久化回调 (bootstrap 接线用)。

    N5 (2026-10-11): 回调在事件循环内被同步触发, 原子写含 fsync —— 有运行中事件
    循环时改 executor 后台执行 (fire-and-forget, 失败在包装内记日志), 不阻塞事件
    循环; 无事件循环 (同步测试直调) 保持原同步行为。
    """
    import asyncio

    def _write() -> None:
        try:
            _persist_links(bus, path)
        except Exception as exc:  # noqa: BLE001 与 bus._trigger_persist 口径一致
            logger.warning("Link 持久化失败, in-memory 状态已变更", error=str(exc))

    def _callback() -> None:
        try:
            loop = asyncio.get_running_loop()
        except RuntimeError:
            _write()
            return
        loop.run_in_executor(None, _write)

    return _callback